requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
python-dotenv>=1.0.0
email-validator>=2.1.0
textblob>=0.17.1
flask>=2.0.0
